some improvements

+++ core/node/node_sync/src/external_io.rs
@@ -353,7 +353,7 @@ impl StateKeeperIO for ExternalIO {
         }
     }
 
-    fn update_next_l2_block_timestamp(&mut self, _block_timestamp: &mut u128) {}
+    fn update_next_l2_block_timestamp(&mut self, _block_timestamp: &mut u64) {}
 
     async fn wait_for_next_tx(
         &mut self,
