+++ core/node/state_keeper/src/io/persistence.rs
@@ -482,7 +482,7 @@ mod tests {
             &default_system_env(),
             Default::default(),
             Default::default(),
-            u128::from(l1_batch_env.first_l2_block.timestamp) * 1000,
+            l1_batch_env.first_l2_block.timestamp * 1000,
         );
         pool.connection()
             .await
